Moves handler funcs into separate file for readability
